+++ suma_multiple.c
@@ -4,12 +4,12 @@
 
 #include "isaac.h"
 
-void suma_multiple(double multiple, int p, int fila_inici, int fila_desti, int cols, double m[][cols]){
-    double aux_vector[cols];
+void suma_multiple(int multiple, int p, int fila_inici, int fila_desti, int cols, int m[][cols]){
+    int aux_vector[cols];
 
     for(int j = 0; j < cols; j++){
         aux_vector[j] = multiple * m[fila_inici][j];
-        m[fila_desti][j] = (int)(m[fila_desti][j] + aux_vector[j]) % p;
+        m[fila_desti][j] = (m[fila_desti][j] + aux_vector[j]) % p;
     }
     return;
 }
