+++ vectorxescalar.c
@@ -4,10 +4,10 @@
 
 #include "isaac.h"
 
-void vectorxescalar(int dim, int a, int p, double vector[]){
-    double resul[dim];
+void vectorxescalar(int dim, int a, int p, int vector[]){
+    int resul[dim];
     for(int i = 0; i < dim; i++){
-        resul[i] = (int)(a * vector[i]) % p;
+        resul[i] = (a * vector[i]) % p;
     }
     printf("El resultat de multiplicar el vector per %d en F(%d) és\n", a, p);
     imprimeixvector(dim, resul);
